requirements.txt
PyYAML>=6.0
requests>=2.28
python-dotenv>=1.0
GitPython>=3.1
openai>=1.0
flask>=2.2
jira>=3.3.1
pytest>=7.0
flake8>=6.0
